[refactor]: #43 rename script.js to main.js
[body]

[notes]

close #<issue-number>
