+++ PongChat/pingpong/templates/pingpong/index.html
@@ -42,7 +42,7 @@
         </button>
       </div>
     </div>
-    <script type="module" src="{% static 'pingpong/js/script.js' %}"></script>
+    <script type="module" src="{% static 'pingpong/js/main.js' %}"></script>
     <script
       src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"
       integrity="sha384-geWF76RCwLtnZ8qwWowPQNguL3RmwHVBC9FhGdlKrxdiJJigb/j/68SIy3Te4Bkz"
